new validations

+++ myphp.php
@@ -17,33 +17,69 @@
 
 // get emloyee by id
  if(isset($_POST["empId"])&&isset($_POST["getEmp"])){
+     
     $stmt = $pdo->prepare('SELECT * FROM __employee WHERE emp_id=:id');
     $stmt->execute(['id' =>$_POST["empId"]]);
     $emp=$stmt->fetch();
+    if($emp['emp_id']==$_POST["empId"]){
 
     echo $emp['emp_name'].$emp['start_date'];
 
     }
+    else{
+        echo "no id found";
+    }
+ }
     // add employee to the table
 
+
  else if(isset($_POST["add"])&&isset($_POST["name"])&&isset($_POST["date"])){
+         $stmt = $pdo->prepare('SELECT * FROM __employee WHERE emp_name=:name');
+         $stmt->execute(['name' =>$_POST["name"]]);
+         $emp=$stmt->fetch();
+       if($emp['emp_name']!=$_POST["name"]){
+
         $statement = $pdo->prepare("INSERT INTO __employee(emp_name,start_date)
                     VALUES(:fname, :adte)");
                     $statement->execute(array(
                     "fname" => $_POST["name"],
-                    "adte" => $_POST["date"]));
+                   "adte" => $_POST["date"]));
+       }
+        else{
+            echo "name already exsits";
+        } 
     }
     //delete by id
+
   else if(isset($_POST["dlt"])&&isset($_POST["empId"])){
+
+       $stmt = $pdo->prepare('SELECT * FROM __employee WHERE emp_id=:id');
+         $stmt->execute(['id' =>$_POST["empId"]]);
+         $emp=$stmt->fetch();
+         if($emp['emp_id']==$_POST["empId"]){
+
          $stmt = $pdo->prepare('DELETE  FROM  __employee  WHERE emp_id =:id');
           $stmt->execute(['id' =>$_POST["empId"]]);
-           
+           }
+         else{
+              echo "no id found";
+         } 
        }
     //update employee by id
-  else if(isset($_POST["update"])&&isset($_POST["empId"])){
-         $stmt = $pdo->prepare('UPDATE __employee  SET emp_name = :name, start_date = :dat1  WHERE emp_id =:id');
-          $stmt->execute(['id' =>$_POST["empId"],'dat1' =>$_POST["date"],'name' =>$_POST["name"]]);
 
+  else if(isset($_POST["update"])&&isset($_POST["empId"])){
+
+         $stmt = $pdo->prepare('SELECT * FROM __employee WHERE emp_id=:id');
+         $stmt->execute(['id' =>$_POST["empId"]]);
+         $emp=$stmt->fetch();
+         if($emp['emp_id']==$_POST["empId"]){
+
+            $stmt = $pdo->prepare('UPDATE __employee  SET emp_name = :name, start_date = :dat1  WHERE emp_id =:id');
+            $stmt->execute(['id' =>$_POST["empId"],'dat1' =>$_POST["date"],'name' =>$_POST["name"]]);
+         }
+         else{
+              echo "no id found";
+         }
         }
     //get all employees
   else if(isset($_POST["all"])){
@@ -51,7 +87,7 @@
           
         foreach ($stmt as $row)
             {
-                echo "name: " .$row['emp_name']." start date: ".$row['start_date']."<br>";
+                echo"id: ".$row['emp_id']. ", name: " .$row['emp_name'].", start date: ".$row['start_date']."<br>";
             }
 
    //error in connection with html
